* `local:docker` runner: this runner takes a test plan docker image and schedules as many containers as instances the user wants.
  * It creates a `bridge` network and also instantiates a redis container if one is needed (for the sync service).
  * It attaches the redis instance and all test containers to the bridge (TODO: we need to dissolve the bridge network once we're done -- currently not doing this).
  * This runner has various options: tail, unstarted, etc. that can be used to tune the behaviour when invoking the run.

* Solidify how builder/runner configuration is managed, and how configurations defined by plan manifests can be overridden by the user at runtime.

* Refactor test case outputs.

* Initialize Redis client from the environment.

* Logging improvements.

+++ pkg/build/docker_go.go
@@ -1,17 +1,19 @@
 package build
 
 import (
-	"bufio"
 	"context"
+	"encoding/json"
 	"fmt"
+	"io"
 	"io/ioutil"
 	"os"
 	"os/exec"
 	"path/filepath"
+	"reflect"
 	"strings"
 	"time"
 
-	"github.com/ipfs/testground/pkg/api"
+	"github.com/docker/docker/pkg/jsonmessage"
 
 	"github.com/docker/docker/api/types"
 	"github.com/docker/docker/api/types/filters"
@@ -22,17 +24,28 @@ import (
 	"github.com/otiai10/copy"
 )
 
+var (
+	_ Builder = &DockerGoBuilder{}
+)
+
 // DockerGoBuilder builds the test plan as a go-based container.
 type DockerGoBuilder struct{}
 
-var _ Builder = (*DockerGoBuilder)(nil)
+type DockerGoBuilderConfig struct {
+	Enabled    bool
+	GoVersion  string `toml:"go_version" overridable:"yes"`
+	ModulePath string `toml:"module_path" overridable:"yes"`
+	ExecPkg    string `toml:"exec_pkg" overridable:"yes"`
+	FreshGomod bool   `toml:"fresh_gomod" overridable:"yes"`
+}
 
 // TODO cache build outputs https://github.com/ipfs/testground/issues/36
 // Build builds a testplan written in Go into a Docker container.
-func (b *DockerGoBuilder) Build(opts *Input, config interface{}) (*Output, error) {
-	cfg, ok := config.(api.GoBuildStrategy)
+func (b *DockerGoBuilder) Build(opts *Input) (*Output, error) {
+	// TODO apply configuration overrides.
+	cfg, ok := opts.BuildConfig.(*DockerGoBuilderConfig)
 	if !ok {
-		panic("expected configuration type GoBuildStrategy")
+		return nil, fmt.Errorf("expected configuration type DockerGoBuilderConfig, was: %T", opts.BuildConfig)
 	}
 
 	var (
@@ -92,21 +105,23 @@ func (b *DockerGoBuilder) Build(opts *Input, config interface{}) (*Output, error
 		return nil, err
 	}
 
-	for _, f := range []string{"go.mod", "go.sum"} {
-		file := filepath.Join(plandst, f)
-		if _, err := os.Stat(file); !os.IsNotExist(err) {
-			if err := os.Remove(file); err != nil {
-				return nil, fmt.Errorf("cleanup failed; %w", err)
+	if cfg.FreshGomod {
+		for _, f := range []string{"go.mod", "go.sum"} {
+			file := filepath.Join(plandst, f)
+			if _, err := os.Stat(file); !os.IsNotExist(err) {
+				if err := os.Remove(file); err != nil {
+					return nil, fmt.Errorf("cleanup failed; %w", err)
+				}
 			}
 		}
-	}
 
-	// Initialize a fresh go.mod file.
-	cmd := exec.Command("go", "mod", "init", cfg.ModulePath)
-	cmd.Dir = plandst
-	out, err := cmd.CombinedOutput()
-	if !strings.Contains(string(out), "creating new go.mod") {
-		return nil, fmt.Errorf("unable to create go.mod; %s", out)
+		// Initialize a fresh go.mod file.
+		cmd := exec.Command("go", "mod", "init", cfg.ModulePath)
+		cmd.Dir = plandst
+		out, _ := cmd.CombinedOutput()
+		if !strings.Contains(string(out), "creating new go.mod") {
+			return nil, fmt.Errorf("unable to create go.mod; %s", out)
+		}
 	}
 
 	// If we have version overrides, apply them.
@@ -123,27 +138,26 @@ func (b *DockerGoBuilder) Build(opts *Input, config interface{}) (*Output, error
 		fmt.Sprintf("-replace=github.com/ipfs/testground/sdk/runtime=../sdk/runtime"))
 
 	// Write replace directives.
-	cmd = exec.Command("go", append([]string{"mod", "edit"}, replaces...)...)
+	cmd := exec.Command("go", append([]string{"mod", "edit"}, replaces...)...)
 	cmd.Dir = plandst
-	out, err = cmd.CombinedOutput()
+	_, err = cmd.CombinedOutput()
 	if err != nil {
 		return nil, fmt.Errorf("unable to add replace directives to go.mod; %w", err)
 	}
-
-	fmt.Println(tmp)
 
 	tar, err := archive.TarWithOptions(tmp, &archive.TarOptions{})
 	if err != nil {
 		return nil, err
 	}
 
+	args := map[string]*string{
+		"GO_VERSION":        &cfg.GoVersion,
+		"TESTPLAN_EXEC_PKG": &cfg.ExecPkg,
+	}
+
 	buildOpts := types.ImageBuildOptions{
-		Dockerfile: "Dockerfile",
-		Tags:       []string{id},
-		BuildArgs: map[string]*string{
-			"GO_VERSION":        &cfg.GoVersion,
-			"TESTPLAN_EXEC_PKG": &cfg.ExecPkg,
-		},
+		Tags:      []string{id},
+		BuildArgs: args,
 	}
 
 	resp, err := cli.ImageBuild(ctx, tar, buildOpts)
@@ -152,9 +166,20 @@ func (b *DockerGoBuilder) Build(opts *Input, config interface{}) (*Output, error
 	}
 	defer resp.Body.Close()
 
-	scan := bufio.NewScanner(resp.Body)
-	for scan.Scan() {
-		fmt.Println(scan.Text())
+	var msg jsonmessage.JSONMessage
+Loop:
+	for dec := json.NewDecoder(resp.Body); ; {
+		switch err := dec.Decode(&msg); err {
+		case nil:
+			msg.Display(os.Stdout, true)
+			if msg.Error != nil {
+				return nil, msg.Error
+			}
+		case io.EOF:
+			break Loop
+		default:
+			return nil, err
+		}
 	}
 
 	return &Output{ArtifactPath: id}, nil
@@ -204,4 +229,12 @@ func imageExists(ctx context.Context, cli *client.Client, id string) (bool, erro
 		return false, err
 	}
 	return len(summary) > 0, nil
+}
+
+func (*DockerGoBuilder) ID() string {
+	return "docker:go"
+}
+
+func (*DockerGoBuilder) ConfigType() reflect.Type {
+	return reflect.TypeOf(DockerGoBuilderConfig{})
 }
